examples/rust-matter-light: Move all "matter core" related stuff into maikerlab/rust-riot-wrappers (branch=with_matter)
